src/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
pandas==2.1.3
pyyaml==6.0.1
lxml==4.9.3
pyppeteer>=1.0.2
aiohttp==3.9.1
playwright==1.40.0
pymongo==4.6.0
sqlalchemy==2.0.23
tqdm==4.66.1
python-dotenv==1.0.0
urllib3>=1.26.0,<3.0.0
webdriver-manager==4.0.1
tenacity==8.2.3
loguru==0.7.2
streamlit==1.45.0
pytube==15.0.0
yt-dlp==2023.11.16
Pillow==10.1.0
